+++ NEWS.md
@@ -1,5 +1,7 @@
 # cvms 0.0.0.9000
 
+* Adds optional parallelization.
+
 * Results now contain a count of singular fit messages. See ?lme4::isSingular for more information.
 
 * Argument "positive" changes default to 2. Now takes either 1 or 2 (previously 0 and 1). If your dependent variable has
@@ -8,6 +10,8 @@ values 0 and 1, 1 is now the positive class by default.
 * AUC calculation has changed. Now explicitly sets the direction in pROC::roc.
 
 * Unit tests have been updated for the new random sampling generator in R 3.6.0. They will NOT run previous versions of R. 
+
+* Adds baseline() for creating baseline evaluations.
 
 * Adds reconstruct_formulas() for reconstructing formulas based on model definition columns in the results tibble.